Move RunContext to dsl.models
